refactor: improve error handling patterns and code quality
- replace map_err with inspect_err for side-effect error logging
- simplify conditional unwrapping using if-let patterns
- fix integer cast warning in status display calculation
- update build command documentation to include clippy

+++ examples/staging_operations.rs
@@ -259,7 +259,7 @@ fn display_status_changes(before: &rustic_git::GitStatus, after: &rustic_git::Gi
         let after_val = after_counts.get(status).unwrap_or(&0);
         
         if before_val != after_val {
-            println!("      {}: {} → {} ({:+})", status, before_val, after_val, *after_val as i32 - *before_val as i32);
+            println!("      {}: {} → {} ({:+})", status, before_val, after_val, *after_val - *before_val);
         }
     }
 }
